full blog park-filter, search, and search+filter functionality completed
